% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC_1_Gauss.R
\name{aic_1_gauss}
\alias{aic_1_gauss}
\title{MLE and AIC from normally distributed observations}
\usage{
aic_1_gauss(x)
}
\arguments{
\item{x}{a numeric vector of observations}
}
\value{
a list with AIC and parameters
}
\description{
MLE and AIC from normally distributed observations
}
\examples{
x <- rnorm(10, m=5, sd=2)
res <- aic_1_gauss(x)

res$mu
res$sigma
res$loglikelihood
sum(dnorm(x, m = res$mu, sd = res$sigma, log = TRUE))
}
\author{
Anthony Devaux, Boris Hejblum
}
