% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC_3_Gauss.R
\name{aic_3_gauss}
\alias{aic_3_gauss}
\title{#' MLE and AIC from mixture of 3 normally distributed observations}
\usage{
aic_3_gauss(x, init, maxit = 15)
}
\arguments{
\item{x}{Numeric vector of observations}

\item{init}{Numeric vector of parameters (p, mu1, mu2, sigma1, sigma2)}
}
\value{
List with AIC and parameters
}
\description{
#' MLE and AIC from mixture of 3 normally distributed observations
}
\examples{
xx <- c(rnorm(1000,4,2), rnorm(2000, -25, 0.1), rnorm(500, 30, 4))
100/350; 200/350; 50/350
aic_3_gauss(xx, init = c(0.3, 0.5, 0, 0, 0, 2, 2, 2))

}
\author{
Boris Hejblum, Anthony Devaux
}
