% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsinhTransformation.R
\name{AsinhTransformation}
\alias{AsinhTransformation}
\title{Data transformation using asinh}
\usage{
AsinhTransformation(M, num_col, a = 0, b = 0.2, c = 0)
}
\description{
Data transformation using asinh
}
\keyword{internal}
