% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryTree.R
\name{BinaryTree}
\alias{BinaryTree}
\title{Builds a binary tree.}
\usage{
BinaryTree(M, minleaf = 1, t = 0.1, verbose = TRUE, force_first_markers = NULL)
}
\description{
Builds a binary tree.
}
\keyword{internal}
