% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapCI.R
\name{bootstrapCI}
\alias{bootstrapCI}
\title{Bootstrapped Confidence Interval.}
\usage{
bootstrapCI(stat, n, alpha)
}
\arguments{
\item{stat}{A numeric vector of statistics for which to compute a 
confidence interval.}

\item{n}{An integer giving the number of bootstrap samples.}

\item{alpha}{A real number comprised in ]0, 1[ : 
1 - desired confidence level.}
}
\description{
Bootstrapped Confidence Interval.
}
\author{
Chariff Alkhassim
}
