% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytofTree.R
\name{CytofTree}
\alias{CytofTree}
\title{Binary tree algorithm for cytometry data analysis.}
\usage{
CytofTree(M, minleaf = 1, t = 0.1, verbose = TRUE,
  force_first_markers = NULL, transformation = c("asinh", "none"),
  num_col = NULL)
}
\arguments{
\item{M}{A matrix of size n x p containing cytometry measures
of n cells on p markers.}

\item{minleaf}{An integer indicating the minimum number of cells
per population. Default is \code{1}.}

\item{t}{A real positive-or-null number used for comparison with
the normalized AIC computed at each node of the tree.
A higher value limits the height of the tree.}

\item{verbose}{A logical controlling if a text progress bar is displayed 
during the execution of the algorithm. By default is TRUE.}

\item{force_first_markers}{a vector of index to split the data on first.
This argument is used in the semi-supervised setting, forcing the algorithm to consider 
those markers first, in the order they appear in this \code{force_first_markers} vector, 
and forcing the split at every node. Default is \code{NULL}, in which case
the clustering algorithm is unsupervised.}
}
\value{
An object of class 'cytomeTree' providing a partitioning
of the set of n cells.
\itemize{
\item{\code{annotation}}{ A \code{data.frame} containing the annotation of each 
cell population underlying the tree pattern.}
\item{\code{labels}}{ The partitioning of the set of n cells.}
\item{\code{M}}{ The input matrix.}
\item{\code{mark_tree}}{ A two level list containing markers used
for node splitting.}
}
}
\description{
Binary tree algorithm for cytometry data analysis.
}
\details{
The algorithm is based on the construction of a binary tree,
the nodes of which are subpopulations of cells. At each node,
observed cells and markers are modeled by both a family of normal
distributions and a family of bi-modal normal mixture distributions.
Splitting is done according to a normalized difference of AIC between
the two families.
}
\examples{
head(DLBCL)
}
\author{
Chariff Alkhassim, Boris Hejblum
}
