% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytomeTree.R
\name{CytomeTree}
\alias{CytomeTree}
\title{Binary tree algorithm for cytometry data analysis.}
\usage{
CytomeTree(M, minleaf = 1, D = 0.1)
}
\arguments{
\item{M}{A matrix of size n x p containing cytometry measures
of n cells on p markers.}

\item{minleaf}{An integer indicating the minimum number of cell
per population. Default is \code{1}.}

\item{D}{A real positive-or-null number used for comparison with
the normalized AIC computed at each node of the tree.
A higher values limits the height of the tree.}
}
\value{
An object of class 'cytomeTree' providing a partitioning
of the set of n cells. 
\itemize{
\item{\code{combinations}}{ A matrix of size n x p containing 
the annotation of each cell given by the tree.}
\item{\code{labels}}{ The partitioning of the set of n cells.}
\item{\code{M}}{ The input matrix.}
\item{\code{mark_tree}}{ A two level list containing markers used 
for node splitting.}
}
}
\description{
Binary tree algorithm for cytometry data analysis.
}
\details{
The algorithm is based on the construction of a binary tree, 
the nodes of which are subpopulations of cells. At each node, 
observed cells and markers are modeled by both a familly of normal
and a familly of normal mixtures distributions. 
Spliting is done according to a normalized difference of AIC between 
the two families.
}
\author{
Chariff Alkhassim
}

