% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLBCL.r
\docType{data}
\name{DLBCL}
\alias{DLBCL}
\title{Diffuse large B-cell lymphoma data set from the FlowCAP-I challenge.}
\format{A data frame with 5524 cell events and 3 markers.}
\source{
\url{http://flowcap.flowsite.org/}
}
\usage{
data(DLBCL)
}
\description{
Diffuse large B-cell lymphoma data set from the FlowCAP-I challenge.
}
\keyword{data}
