% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Fmeasure_costC_arma}
\alias{Fmeasure_costC_arma}
\title{C++ implementation of cost computation with Fmeasure as loss function
using the Armadillo library}
\usage{
Fmeasure_costC_arma(c)
}
\arguments{
\item{c}{matrix where each column is one partition}
}
\description{
C++ implementation of cost computation with Fmeasure as loss function
using the Armadillo library
}

