% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FmeasureC_no0}
\alias{FmeasureC_no0}
\title{C++ implementation of the F-measure computation without the reference class labeled "0"}
\usage{
FmeasureC_no0(pred, ref)
}
\arguments{
\item{pred}{vector of a predicted partition}

\item{ref}{vector of a reference partition}
}
\description{
Aghaeepour in FlowCAP 1 ignore the reference class labeled "0"
}
\author{
Boris Hejblum
}
