% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussMix2.r
\name{GaussMix2}
\alias{GaussMix2}
\title{tri-modal normal mixture distribution.}
\usage{
GaussMix2(x, mu1, mu2, mu3, sd1, sd2, sd3, pi1, pi2, pi3)
}
\description{
tri-modal normal mixture distribution.
}
\keyword{internal}
