% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gridsearch.R
\name{Gridsearch}
\alias{Gridsearch}
\title{Gridsearch}
\usage{
Gridsearch(x, resu_aic_1_gauss, iter_max = 15, ntry = 30,
  mixture = 2)
}
\description{
Gridsearch
}
\examples{
x <- c(rnorm(100,4,2), rnorm(200, -25, 0.1))
resu_aic_1_gauss <- aic_1_gauss(x)
res <- Gridsearch(x, resu_aic_1_gauss)
}
