% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIPC.r
\docType{data}
\name{HIPC}
\alias{HIPC}
\title{HIPC T cell data set from HIPC program, patient 12828. The data was analyzed and gated by Stanford.}
\format{A data frame with 33992 cell events and 6 markers.}
\source{
\url{https://www.immuneprofiling.org/} \url{https://www.immunespace.org/} 
\url{https://www.immunespace.org/project/HIPC/Lyoplate/begin.view?pageId=study.DATA_ANALYSIS}
}
\usage{
data(HIPC)
}
\description{
HIPC T cell data set from HIPC program, patient 12828. The data was analyzed and gated by Stanford.
}
\keyword{data}
