% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMdata.R
\docType{data}
\name{IMdata}
\alias{IMdata}
\title{Influenza vaccine response dataset}
\format{A data frame with 10,000 rows and 39 variables:}
\source{
\url{https://www.immunespace.org/project/Studies/SDY478/begin.view?}
}
\usage{
data(IMdata)
}
\description{
A dataset containing 10,000 cells and 39 markers of mass cytometry subsampled
from the sample SUB116516.478 from the study SDY478 by Mark Davis 
retrieved from ImmuneSpace
}
\keyword{data}
