% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KmeansOPT.R
\name{KmeansOPT}
\alias{KmeansOPT}
\title{Finds the partition which minimize the within-leaves sum of squares.}
\usage{
KmeansOPT(groups, leaves, labels, dat, K)
}
\description{
Finds the partition which minimize the within-leaves sum of squares.
}
\keyword{internal}
