% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeavesCenters.R
\name{LeavesCenters}
\alias{LeavesCenters}
\title{Computes means of leaves given cytomeTreeObj.}
\usage{
LeavesCenters(cytomeTreeObj)
}
\description{
Computes means of leaves given cytomeTreeObj.
}
\keyword{internal}
