% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeavesMedians.R
\name{LeavesMedians}
\alias{LeavesMedians}
\title{Computes medians of leaves given cytomeTreeObj.}
\usage{
LeavesMedians(cytomeTreeObj)
}
\description{
Computes medians of leaves given cytomeTreeObj.
}
\keyword{internal}
