% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log10Transformation.R
\name{Log10Transformation}
\alias{Log10Transformation}
\title{Data transformation using log10}
\usage{
Log10Transformation(M, num_col, a = 1, b = 1)
}
\description{
Data transformation using log10
}
\keyword{internal}
