% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.r
\name{plot_graph}
\alias{plot_graph}
\title{Plot the binary tree built using CytomeTree.}
\usage{
plot_graph(CytomeTreeObj, Ecex = 1, Ecolor = 8, Vcex = 0.8, Vcolor = 0, ...)
}
\arguments{
\item{CytomeTreeObj}{An object of class CytomeTree.}

\item{Ecex}{Number indicating the amount by which text
on the edges should be scaled. Default is \code{1}.}

\item{Ecolor}{An integer or a string of character
to color edges of the graph.  Default is \code{8}.}

\item{Vcex}{Number indicating the amount by which text
in the vertices should be scaled.  Default is \code{.8}.}

\item{Vcolor}{A vector of class numeric or character to color
vertices of the graph. Default is \code{0}.}

\item{...}{additional arguments to be passed to \code{\link{plot_graph}}}
}
\description{
Plot the binary tree built using CytomeTree.
}
\author{
Chariff Alkhassim
}
