% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_HMpop.R
\name{plot_HMpop}
\alias{plot_HMpop}
\title{Plot a heatmap representing the median of the expression for each population and marker using CytomeTree}
\usage{
plot_HMpop(TreeObj, AnnotObj)
}
\arguments{
\item{TreeObj}{An object of class CytomeTree}

\item{AnnotObj}{An object of class Annotation}
}
\description{
Plot a heatmap representing the median of the expression for each population and marker using CytomeTree
}
\examples{

# Run CytomeTree
data(DLBCL)
cellevents <- DLBCL[,c("FL1", "FL2", "FL4")]
Tree <- CytomeTree(cellevents, minleaf = 1, t=.1)
Annot <- Annotation(Tree,plot=FALSE)

# Plot the cell count
plot_HMpop(Tree,Annot)
}
\author{
Anthony Devaux, Boris Hejblum
}
