% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nodes.R
\name{plot_nodes}
\alias{plot_nodes}
\title{Plot the distribution of the observed cells at each node 
of the binary tree built using CytomeTree.}
\usage{
plot_nodes(CytomeTreeObj, nodes = NULL)
}
\arguments{
\item{CytomeTreeObj}{An object of class CytomeTree.}

\item{nodes}{A vector of class character containing the name of
the nodes for which the distribution is to be plotted. Default is 
\code{NULL}, and plots the distribution of each node.}
}
\description{
Plot the distribution of the observed cells at each node 
of the binary tree built using CytomeTree.
}
\author{
Chariff Alkhassim
}
