% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetrievePops.r
\name{RetrievePops}
\alias{RetrievePops}
\title{Retrieve cell populations found using Annotation.}
\usage{
RetrievePops(AnnotationObj, phenotypes)
}
\arguments{
\item{AnnotationObj}{An object of class Annotation.}

\item{phenotypes}{A list containing at least one element of class 
matrix describing a sought phenotype. Each matrix should have two
columns where the name of a used marker is associated to a value
chosen between 0, 1 and 2. 0 translates to -, 1 to + and 2 to ++.}
}
\value{
A \code{list} of two elements.
\itemize{
\item{\code{phenotypesinfo}}{ A \code{list} containing informations 
about sought populations.}
\item{\code{Mergedleaves}}{ The partitioning of the set of n cells
with potentially merged leaves.}
}
}
\description{
Retrieve cell populations found using Annotation.
}
\author{
Chariff Alkhassim, Boris Hejblum
}
