% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC_2_Gauss.R
\name{aic_2_gauss}
\alias{aic_2_gauss}
\title{Compute parameters and AIC from observations with mixture of 2 normal distribution}
\usage{
aic_2_gauss(x, init, maxit = 15)
}
\arguments{
\item{x}{Numeric vector of observations}

\item{init}{Numeric vector of parameters (p, mu1, mu2, sigma1, sigma2)}
}
\value{
List with AIC and parameters
}
\description{
Compute parameters and AIC from observations with mixture of 2 normal distribution
}
\examples{
xx <- c(rnorm(100,4,2), rnorm(200, -25, 0.1))
aic_2_gauss(xx, init = c(0.5, 0, 0, 2, 2))

}
\author{
Boris Hejblum, Anthony Devaux
}
