% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{Annotation}
\alias{Annotation}
\title{Annotates cell populations found using CytomeTree.}
\usage{
Annotation(CytomeTreeObj, K3markers = NULL, plot = TRUE)
}
\arguments{
\item{CytomeTreeObj}{An object of class CytomeTree.}

\item{K3markers}{A vector of class character where the names of 
the markers for which 3 levels of expression are seeked can be specified.
Default is \code{NULL} i.e. 2 levels of expression per marker.}

\item{plot}{A logical value indicating whether or not to plot the 
partitioning in 2 or 3 groups for each marker. Default is \code{TRUE}.}
}
\value{
A \code{data.frame} containing the annotation of each 
cell population.
}
\description{
Annotates cell populations found using CytomeTree.
}
\details{
The algorithm is set to find the partitioning in 2 or
3 groups of cell populations found using CytomeTree. It minimize 
the within-leaves sum of squares of the observed values on each
marker.
}
\author{
Chariff Alkhassim
}

