% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{DLBCL}
\alias{DLBCL}
\title{Diffuse large B-cell lymphoma data set from the flowcap challenge.}
\format{A data frame with 5524 cell events and 3 markers.}
\source{
\url{http://http://flowcap.flowsite.org/}
}
\usage{
DLBCL
}
\description{
Diffuse large B-cell lymphoma data set from the flowcap challenge.
}
\keyword{datasets}
