% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIPC.r
\docType{data}
\name{HIPC}
\alias{HIPC}
\title{HIPC T cell panel data set from HIPC program, patient 12828. The data was analyzed and gated by Stanford.}
\format{
A data frame with 33992 cell events and 6 markers.
}
\source{
\url{https://www.immuneprofiling.org/} \url{https://www.immunespace.org/} 
\url{https://www.immunespace.org/project/HIPC/Lyoplate/begin.view?pageId=study.DATA_ANALYSIS}
}
\usage{
data(HIPC)
}
\description{
HIPC T cell panel data set from HIPC program, patient 12828. The data was analyzed and gated by Stanford.
}
\details{
This immunophenotyping T cell panel from the Lyoplate HIPC dataset 
was used as part of the FlowCAP III Lyoplate challenge.
}
\references{
Maecker HT, McCoy JP & Nussenblatt R (2012). Standardizing 
immunophenotyping for the human immunology project. Nature Reviews Immunology, 
12(3):191--200. DOI: 10.1038/nri3158

Finak G, Langweiler M, Jaimes M, Malek M, Taghiyar J, Korin Y, Raddassi K, 
Devine L, Obermoser G, Pekalski ML, Pontikos N, Diaz A, Heck S, Villanova F, 
Terrazzini N, Kern F, Qian Y, Stanton R, Wang K, Brandes A, Ramey J, 
Aghaeepour N, Mosmann T, Scheuermann RH, Reed E, Palucka K, Pascual V, 
Blomberg BB, Nestle F, Nussenblatt RB, Brinkman RR, Gottardo R, Maecker H & 
McCoy JP (2016). Standardizing Flow Cytometry Immunophenotyping Analysis from the Human 
ImmunoPhenotyping Consortium. Scientific Reports. 10(6):20686. DOI: 10.1038/srep20686.
}
\keyword{data}
