% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partition2gr.r
\name{Partition2gr}
\alias{Partition2gr}
\title{Computes N-1 possible partitions of 2 sets of N.}
\usage{
Partition2gr(N)
}
\description{
Computes N-1 possible partitions of 2 sets of N.
}
\keyword{internal}

