% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partition3gr.r
\name{Partition3gr}
\alias{Partition3gr}
\title{Computes (N-1)*(N-2) possible partitions of 3 sets of N.}
\usage{
Partition3gr(N)
}
\description{
Computes (N-1)*(N-2) possible partitions of 3 sets of N.
}
\keyword{internal}

