% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeAnnot.R
\name{TreeAnnot}
\alias{TreeAnnot}
\title{Returns the underlying annotation given a tree pattern.}
\usage{
TreeAnnot(labels, combinaisons)
}
\description{
Returns the underlying annotation given a tree pattern.
}
\keyword{internal}
